Normalize files
